+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Changelog
 
+## [0.2.1](https://github.com/Blobfolio/oxford_join/releases/tag/v0.2.1) - 2022-01-13
+
+### Added
+
+* `Conjunction::is_empty`
+* impl `Hash`, `Borrow<str>` for `Conjunction`
+
+
+
 ## [0.2.0](https://github.com/Blobfolio/oxford_join/releases/tag/v0.2.0) - 2021-10-21
 
 ### Added
 
-- This changelog! Haha.
+* This changelog! Haha.
 
 ### Changed
 
-- Use Rust edition 2021.
+* Use Rust edition 2021.